+++ Sources/KeyboardManager/KeyboardManager.swift
@@ -42,8 +42,16 @@ open class KeyboardManager: NSObject, UIGestureRecognizerDelegate {
     /// A flag that indicates if a portion of the keyboard is visible on the screen
     private(set) public var isKeyboardHidden: Bool = true
     
+    /// A flag that indicates if the additional bottom space should be applied to
+    /// the interactive dismissal of the keyboard
+    public var shouldApplyAdditionBottomSpaceToInteractiveDismissal: Bool = false
+
     // MARK: - Properties [Private]
     
+    /// The additional bottom space specified for laying out the input accessory view
+    /// when binding to it
+    private var additionalBottomSpace: (() -> CGFloat)?
+
     /// The `NSLayoutConstraintSet` that holds the `inputAccessoryView` to the bottom if its superview
     private var constraints: NSLayoutConstraintSet?
     
@@ -143,6 +151,7 @@ open class KeyboardManager: NSObject, UIGestureRecognizerDelegate {
             fatalError("`inputAccessoryView` must have a superview")
         }
         self.inputAccessoryView = inputAccessoryView
+        self.additionalBottomSpace = additionalBottomSpace
         inputAccessoryView.translatesAutoresizingMaskIntoConstraints = false
         constraints = NSLayoutConstraintSet(
             bottom: inputAccessoryView.bottomAnchor.constraint(equalTo: superview.bottomAnchor),
@@ -287,7 +296,13 @@ open class KeyboardManager: NSObject, UIGestureRecognizerDelegate {
         frame.size.height = window.bounds.height - frame.origin.y
         keyboardNotification.endFrame = frame
 
-        let yCoordinateDirectlyAboveKeyboard = -frame.height
+        var yCoordinateDirectlyAboveKeyboard = -frame.height
+
+        if shouldApplyAdditionBottomSpaceToInteractiveDismissal,
+           let additionalBottomSpace = additionalBottomSpace {
+            yCoordinateDirectlyAboveKeyboard -= additionalBottomSpace()
+        }
+
         /// If a tab bar is shown, letting this number becoming > 0 makes it so the accessoryview disappears below the tab bar. setting the max value to 0 prevents that
         let aboveKeyboardAndAboveTabBar = min(0, yCoordinateDirectlyAboveKeyboard)
         self.constraints?.bottom?.constant = aboveKeyboardAndAboveTabBar